config/iOS_12_min_deployment (#54)
https://app.asana.com/0/1204463818767003/1209662604683008/f

Bumps minimum iOS version to iOS 12

<!-- Korbit AI PR Description Start -->
## Description by Korbit AI

### What change is being made?
Update the iOS deployment target to version 12.0 in the MAX Unreal
Plugin project settings.

### Why are these changes being made?
The update to iOS 12.0 as the minimum deployment target is necessary to
leverage newer iOS features and capabilities, ensuring better
performance, security, and user experience for the app. Additionally,
support for older iOS versions is becoming less relevant, reducing the
need to maintain backward compatibility with iOS 9.0 and 11.0.

> Is this description stale? Ask me to generate a new description by
commenting `/korbit-generate-pr-description`
<!-- Korbit AI PR Description End -->